Add casting for smartcontract::event::decimal's uint8 to operate with go/proto-defined types, which supports only int32 n above via proto-gen

+++ cmd/ebrelayer/txs/parser.go
@@ -73,9 +73,9 @@ func EthereumEventToEthBridgeClaim(valAddr sdk.ValAddress, event types.EthereumE
 	witnessClaim.CosmosReceiver = recipient.String()
 	witnessClaim.Amount = amount
 	witnessClaim.ClaimType = event.ClaimType
-	witnessClaim.Decimals = event.Decimals
+	witnessClaim.Decimals = int32(event.Decimals)
 	witnessClaim.TokenName = event.Name
-	witnessClaim.DenomHash = ethbridge.GetDenomHash(networkDescriptor, tokenContractAddress.String(), event.Decimals, event.Name, event.Symbol)
+	witnessClaim.DenomHash = ethbridge.GetDenomHash(networkDescriptor, tokenContractAddress.String(), int32(event.Decimals), event.Name, event.Symbol)
 
 	return witnessClaim, nil
 }
